Rename plugin to io.beekeeper.gradle.plugins.formatter
We are renaming this plugin because it is conflicting with Silvios
plugin
