+++ beekeeper-formatter-plugin/src/main/resources/META-INF/gradle-plugins/io.beekeeper.formatting.properties
@@ -1,1 +1,1 @@
-implementation-class=io.beekeeper.formatting.FormatterPlugin
+implementation-class=io.beekeeper.formatter.FormatterPlugin
